feat:화면 중앙에 마커
