+++ app/src/main/AndroidManifest.xml
@@ -22,6 +22,7 @@
         android:theme="@style/Theme.KakaoNavi"
         tools:targetApi="31"
         tools:ignore="GoogleAppIndexingWarning">
+
         <activity android:name=".activity.IntroActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenLayout|uiMode|screenSize|smallestScreenSize|locale|layoutDirection|density"
             android:exported="true"
