+++ app/src/main/java/com/hyundailogics/shupool/activity/IntroActivity.kt
@@ -5,11 +5,13 @@ import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Bundle
+import android.util.Log
 import androidx.appcompat.app.AlertDialog
 import androidx.core.app.ActivityCompat
 import com.hyundailogics.shupool.BuildConfig.*
 import com.hyundailogics.shupool.R
 import com.hyundailogics.shupool.application.GlobalApplication
+import com.kakao.sdk.common.util.Utility
 import com.kakaomobility.knsdk.KNLanguageType
 import com.kakaomobility.knsdk.common.objects.KNError_Code_C302
 import kotlinx.coroutines.CoroutineScope
@@ -27,7 +29,6 @@ class IntroActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_intro)
-
         if (!isTaskRoot()) {
             var intent = getIntent()
             var intentAction = intent.action
