+++ app/src/main/java/com/hyundailogics/shupool/adapter/FragmentSearchAdapter.kt
@@ -1,5 +1,6 @@
 package com.kakaomobility.knsample.adapter
 
+import android.content.Intent
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
