--- app/src/main/res/layout/fragment_map_sample_02.xml
@@ -1,152 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".fragment.FragmentMapSample02">
-
-        <com.kakaomobility.knsdk.map.knmapview.KNMapView
-            android:id="@+id/mapView"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="1.0" />
-
-        <Button
-            android:id="@+id/btn00"
-            android:layout_width="30dp"
-            android:layout_height="30dp"
-            android:layout_marginStart="@dimen/size_5"
-            android:layout_marginTop="@dimen/size_5"
-            android:background="@drawable/circle_btn_selectors"
-            android:text="-"
-            android:textColor="@color/white"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TableLayout
-            android:id="@+id/location_container"
-            android:layout_width="290dp"
-            android:layout_height="wrap_content"
-            android:columnCount="2"
-            android:orientation="vertical"
-            android:rowCount="5"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/btn00">
-
-            <TableRow android:id="@+id/location_row_01">
-
-                <Button
-                    android:id="@+id/btn01"
-                    android:layout_width="@dimen/size_120"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/size_5"
-                    android:layout_marginTop="@dimen/size_5"
-                    android:background="@drawable/sample_button_selector"
-                    android:text="@string/map_sample_02_btn_01"
-                    app:layout_constraintLeft_toLeftOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-                <Button
-                    android:id="@+id/btn02"
-                    android:layout_width="@dimen/size_120"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/size_5"
-                    android:layout_marginTop="@dimen/size_5"
-                    android:background="@drawable/sample_button_selector"
-                    android:text="@string/map_sample_02_btn_02"
-                    app:layout_constraintLeft_toRightOf="@id/btn01"
-                    app:layout_constraintTop_toTopOf="@id/btn01" />
-            </TableRow>
-
-            <TableRow android:id="@+id/location_row_02">
-
-                <Button
-                    android:id="@+id/btn03"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/size_5"
-                    android:layout_marginTop="@dimen/size_5"
-                    android:background="@drawable/sample_button_selector"
-                    android:text="@string/map_sample_02_btn_03"
-                    app:layout_constraintLeft_toLeftOf="@id/btn01"
-                    app:layout_constraintRight_toRightOf="@id/btn01"
-                    app:layout_constraintTop_toBottomOf="@id/btn01" />
-
-                <Button
-                    android:id="@+id/btn04"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/size_5"
-                    android:layout_marginTop="@dimen/size_5"
-                    android:background="@drawable/sample_button_selector"
-                    android:text="@string/map_sample_02_btn_04"
-                    app:layout_constraintLeft_toLeftOf="@id/btn03"
-                    app:layout_constraintRight_toRightOf="@id/btn03"
-                    app:layout_constraintTop_toBottomOf="@id/btn03" />
-            </TableRow>
-
-            <TableRow android:id="@+id/location_row_03">
-
-                <Button
-                    android:id="@+id/btn05"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/size_5"
-                    android:layout_marginTop="@dimen/size_5"
-                    android:background="@drawable/sample_button_selector"
-                    android:text="@string/map_sample_02_btn_05"
-                    app:layout_constraintLeft_toLeftOf="@id/btn04"
-                    app:layout_constraintRight_toRightOf="@id/btn04"
-                    app:layout_constraintTop_toBottomOf="@id/btn04" />
-
-                <Button
-                    android:id="@+id/btn06"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/size_5"
-                    android:layout_marginTop="@dimen/size_5"
-                    android:background="@drawable/sample_button_selector"
-                    android:text="@string/map_sample_02_btn_06"
-                    app:layout_constraintLeft_toLeftOf="@id/btn05"
-                    app:layout_constraintRight_toRightOf="@id/btn05"
-                    app:layout_constraintTop_toBottomOf="@id/btn05" />
-            </TableRow>
-
-            <TableRow android:id="@+id/location_row_04">
-
-                <Button
-                    android:id="@+id/btn07"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/size_5"
-                    android:layout_marginTop="@dimen/size_5"
-                    android:background="@drawable/sample_button_selector"
-                    android:text="@string/map_sample_02_btn_07"
-                    app:layout_constraintLeft_toLeftOf="@id/btn06"
-                    app:layout_constraintRight_toRightOf="@id/btn06"
-                    app:layout_constraintTop_toBottomOf="@id/btn06" />
-
-                <Button
-                    android:id="@+id/btn08"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/size_5"
-                    android:layout_marginTop="@dimen/size_5"
-                    android:background="@drawable/sample_button_selector"
-                    android:text="@string/map_sample_02_btn_08"
-                    app:layout_constraintLeft_toLeftOf="@id/btn07"
-                    app:layout_constraintRight_toRightOf="@id/btn07"
-                    app:layout_constraintTop_toBottomOf="@id/btn07" />
-            </TableRow>
-        </TableLayout>
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</layout>
