+++ app/src/main/res/layout/fragment_search.xml
@@ -13,6 +13,8 @@
         android:layout_height="match_parent"
         android:clipToPadding="false"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+        tools:layout_editor_absoluteX="0dp"
+        tools:layout_editor_absoluteY="0dp"
         tools:listitem="@layout/fragment_search_history_item" />
 
     <androidx.recyclerview.widget.RecyclerView
