+++ app/src/main/res/values/strings.xml
@@ -771,4 +771,5 @@
     <string name="full_route_activity_info">%s, %s</string>
     <string name="full_route_activity_waypoint_dist">거리 %s</string>
     <string name="minus">-</string>
+    <string name="move_marker">marker</string>
 </resources>
